[refactor] test_ssd refactoring test sdd 용 directory 생성 및 상수, 기능별 파일 구분
