+++ ssd_tests/test_ssd.py
@@ -3,67 +3,12 @@ from unittest.mock import DEFAULT
 
 import pytest
 
-
 from ssd import SSD
 from nand import Nand
 
+from test_ssd_constants import *
+from test_ssd_utils import *
 
-LBA_LENGTH = 100
-NAND_FILE = "test_ssd_nand.txt"
-OUTPUT_FILE = "test_ssd_output.txt"
-DEFAULT_DATA = "0x00000000"
-WRITE_COMMAND = "W"
-READ_COMMAND = "R"
-ERASE_COMMAND = "E"
-ERROR_MESSAGE = "ERROR"
-FIRST_ADDRESS = "0"
-LAST_ADDRESS = f"{LBA_LENGTH-1}"
-WRITE_SUCCESS_MESSAGE = ""
-
-def remove_files():
-    if os.path.exists(NAND_FILE):
-        os.remove(NAND_FILE)
-    if os.path.exists(OUTPUT_FILE):
-        os.remove(OUTPUT_FILE)
-
-
-def read_output_file():
-    with open(OUTPUT_FILE, 'r') as f:
-        return f.read()
-
-def assert_output_file(expected):
-    assert read_output_file() == expected
-
-
-def check_write_and_read(_ssd, addr, wdata):
-    _ssd.run([WRITE_COMMAND, addr, wdata])
-    _ssd.run([READ_COMMAND, addr])
-    assert_output_file(wdata)
-
-
-def read_data(_ssd, addr):
-    _ssd.run([READ_COMMAND, addr])
-    return read_output_file()
-
-def write_data(_ssd, addr, data):
-    _ssd.run([WRITE_COMMAND, addr, data])
-
-def erase_data(_ssd, addr, size):
-    _ssd.run([ERASE_COMMAND, addr, size])
-
-@pytest.fixture
-def clean_ssd():
-    remove_files()
-    nand = Nand(LBA_LENGTH, NAND_FILE)
-    SSD.LBA_LENGTH = LBA_LENGTH
-    SSD.OUTPUT_FILE = OUTPUT_FILE
-    return SSD(nand)
-
-@pytest.fixture
-def ssd_and_device(mocker):
-    device = mocker.Mock()
-    ssd = SSD(device)
-    return (ssd, device)
 
 class TestSsdWithMock:
     def test_read_command(self, ssd_and_device):
@@ -146,42 +91,17 @@ class TestSsdWithMock:
         assert ssd.result == ERROR_MESSAGE
 
 
-class TestSsdEraseWithMock:
-    @pytest.mark.parametrize( ("addr", "size"), [
-        (FIRST_ADDRESS, "1"),
-        (LAST_ADDRESS, "1"),
-        (FIRST_ADDRESS, "10")
-    ])
-    def test_erase_1_call(self, ssd_and_device, addr, size):
-        ssd, device = ssd_and_device
-        ssd.run([ERASE_COMMAND, addr, size])
-        device.erase.assert_called_with(int(addr), int(1))
-
-    def test_invalid_erase_size(self, ssd_and_device):
-        ssd, device = ssd_and_device
-        ssd.run([ERASE_COMMAND, FIRST_ADDRESS, "11"])
-        assert ssd.result == ERROR_MESSAGE
-
-    def test_erase_out_of_bounds(self, ssd_and_device):
-        ssd, device = ssd_and_device
-        ssd.run([ERASE_COMMAND, LAST_ADDRESS, "2"])
-        assert ssd.result == ERROR_MESSAGE
-
-    def test_erase_invalid_address(self, ssd_and_device):
-        ssd, device = ssd_and_device
-        ssd.run([ERASE_COMMAND, "-1", "2"])
-        assert ssd.result == ERROR_MESSAGE
 
 
 class TestSsd:
-    def test_file_creation(self, clean_ssd):
+    def test_file_creation(self, ssd):
         """
         SSD 객체 맨 처음 생성 후 file들이 생성 되었는지
         """
         assert os.path.exists(OUTPUT_FILE)
 
     @pytest.mark.skip
-    def test_initial_nand_file(self, clean_ssd):
+    def test_initial_nand_file(self, ssd):
         """
         nand 파일 초기화가 잘 되었는지
         """
@@ -192,7 +112,7 @@ class TestSsd:
             assert first_line == f"{DEFAULT_DATA}"
             assert last_line == f"{DEFAULT_DATA}"
 
-    def test_initial_output_file(self, clean_ssd):
+    def test_initial_output_file(self, ssd):
         """
         output 파일 초기화가 잘 되었는지
         """
@@ -200,30 +120,30 @@ class TestSsd:
             content = f.read()
             assert content == ""
 
-    def test_read_empty_data(self, clean_ssd):
+    def test_read_empty_data(self, ssd):
         """
         0 번 주소, 마지막 주소 read 해서 0x00000000 읽히는 지 확인
         """
-        clean_ssd.run([READ_COMMAND, "0"])
+        ssd.run([READ_COMMAND, "0"])
         assert_output_file(DEFAULT_DATA)
-        clean_ssd.run([READ_COMMAND, f"{LBA_LENGTH-1}"])
+        ssd.run([READ_COMMAND, f"{LBA_LENGTH - 1}"])
         assert_output_file(DEFAULT_DATA)
 
-    def test_write_data(self, clean_ssd):
+    def test_write_data(self, ssd):
         """
         0번 주소, 마지막 주소 write 해서 wdata가 읽히는 지 확인
         """
-        check_write_and_read(clean_ssd, "0", "0x1234ABCD")
-        check_write_and_read(clean_ssd, f"{LBA_LENGTH - 1}", "0xAAAA5555")
+        check_write_and_read(ssd, "0", "0x1234ABCD")
+        check_write_and_read(ssd, f"{LBA_LENGTH - 1}", "0xAAAA5555")
 
-    def test_repeat_write_to_same_address(self, clean_ssd):
+    def test_repeat_write_to_same_address(self, ssd):
         """
         같은 주소에 write 반복
         """
         for wdata in ["0xABCD1234", "0xFFFFFFFF", "0xAAAAAAAA"]:
-            check_write_and_read(clean_ssd, "10", wdata)
+            check_write_and_read(ssd, "10", wdata)
 
-    def test_write_multiple_address(self, clean_ssd):
+    def test_write_multiple_address(self, ssd):
         addr_data_pairs = [
             ("10", "0xAAAA5555"),
             ("20", "0xBBBB4444")
@@ -231,43 +151,25 @@ class TestSsd:
 
         # Write to multiple addresses
         for addr, data in addr_data_pairs:
-            clean_ssd.run([WRITE_COMMAND, addr, data])
+            ssd.run([WRITE_COMMAND, addr, data])
 
         # Read and verify each address
         for addr, expected_data in addr_data_pairs:
-            clean_ssd.run([READ_COMMAND, addr])
+            ssd.run([READ_COMMAND, addr])
             assert_output_file(expected_data)
 
-    def test_write_success(self, clean_ssd):
-        clean_ssd.run([READ_COMMAND, FIRST_ADDRESS])
-        clean_ssd.run([WRITE_COMMAND, FIRST_ADDRESS, DEFAULT_DATA])
+    def test_write_success(self, ssd):
+        ssd.run([READ_COMMAND, FIRST_ADDRESS])
+        ssd.run([WRITE_COMMAND, FIRST_ADDRESS, DEFAULT_DATA])
 
         assert_output_file(WRITE_SUCCESS_MESSAGE)
 
-    def test_out_of_lba_range_read(self, clean_ssd):
-        clean_ssd.run([READ_COMMAND, f"{LBA_LENGTH}"])
+    def test_out_of_lba_range_read(self, ssd):
+        ssd.run([READ_COMMAND, f"{LBA_LENGTH}"])
         assert_output_file(ERROR_MESSAGE)
 
-    def test_out_of_lba_range_write(self, clean_ssd):
-        clean_ssd.run([WRITE_COMMAND, "12345678", "0x000000000"])
+    def test_out_of_lba_range_write(self, ssd):
+        ssd.run([WRITE_COMMAND, "12345678", "0x000000000"])
         assert_output_file(ERROR_MESSAGE)
 
 
-
-class TestSsdErase:
-    def test_erase_1(self, clean_ssd):
-        addr = FIRST_ADDRESS
-        wdata = "0x1234abcd"
-        check_write_and_read(clean_ssd, addr, wdata)
-        erase_data(clean_ssd, FIRST_ADDRESS, "1")
-        assert read_data(clean_ssd, FIRST_ADDRESS) == DEFAULT_DATA
-
-    def test_erase_10(self, clean_ssd):
-        for i in range(10):
-            addr = f"{i}"
-            wdata = f"0x0000000{i}"
-            write_data(clean_ssd, addr, wdata)
-        erase_data(clean_ssd, FIRST_ADDRESS, "10")
-        for i in range(10):
-            addr = f"{i}"
-            assert read_data(clean_ssd, addr) == DEFAULT_DATA
