+++ ssd_tests/test_ssd_constants.py
@@ -1,0 +1,13 @@
+
+
+LBA_LENGTH = 100
+NAND_FILE = "test_ssd_nand.txt"
+OUTPUT_FILE = "test_ssd_output.txt"
+DEFAULT_DATA = "0x00000000"
+WRITE_COMMAND = "W"
+READ_COMMAND = "R"
+ERASE_COMMAND = "E"
+ERROR_MESSAGE = "ERROR"
+FIRST_ADDRESS = "0"
+LAST_ADDRESS = f"{LBA_LENGTH-1}"
+WRITE_SUCCESS_MESSAGE = ""
